move ModuleState to libs/types
