+++ apps/web/types/index.d.ts
@@ -113,14 +113,3 @@ export type ContextMenuProps = {
   menu: ContextMenuType
   flowNode?: GraphFlowNode
 }
-
-/**
- * For module styling
- */
-export type ModuleState = {
-  done: ModuleStatus
-  canTake?: boolean
-  suggested?: boolean
-}
-
-export type ModuleStateDict = Record<string, ModuleState>
