+++ libs/types/src/flow.ts
@@ -1,3 +1,5 @@
+import { ModuleStatus } from './repository'
+
 export type FlowNodeCondensed = {
   title: string
   moduleCode: string
@@ -12,3 +14,14 @@ export type FlowEdgeCondensed = {
   source: string
   target: string
 }
+
+/**
+ * For module styling
+ */
+export type ModuleState = {
+  done: ModuleStatus
+  canTake?: boolean
+  suggested?: boolean
+}
+
+export type ModuleStateDict = Record<string, ModuleState>
